Delete README.md
lsys is a maven project coded in groovy.  

To see it work, open images.html in a browser, then invoke the drawing.lsys.LSystemTest.testLoop() method.  The result should be a wiki fractal plant image in the browser changing its shape due to random angle corrections applied to the drawing.